requirements.txt
pyyaml
sympy
pygame
